+++ src/internal/modal.ts
@@ -9,6 +9,8 @@ export default class Modal {
   tabDirection: 'forward' | 'backward' = 'forward';
   currentFocus: HTMLElement | null;
 
+  private tabbableElements: HTMLElement[] | null = null
+
   constructor(element: HTMLElement) {
     this.element = element;
   }
@@ -48,7 +50,12 @@ export default class Modal {
 
   private checkFocus() {
     if (this.isActive() && !this.isExternalActivated) {
-      const tabbableElements = getTabbableElements(this.element);
+      let tabbableElements = this.tabbableElements
+
+      if (!tabbableElements) {
+        tabbableElements = getTabbableElements(this.element);
+      }
+
       if (!this.element.matches(':focus-within')) {
         const start = tabbableElements[0];
         const end = tabbableElements[tabbableElements.length - 1];
@@ -79,6 +86,7 @@ export default class Modal {
     event.preventDefault();
 
     const tabbableElements = getTabbableElements(this.element);
+    this.tabbableElements = tabbableElements
 
     // Because sometimes focus can actually be taken over from outside sources,
     // we don't want to rely on `this.currentFocus`. Instead we check the actual `activeElement` and
@@ -88,7 +96,8 @@ export default class Modal {
 
     if (currentFocusIndex === -1) {
       this.currentFocus = tabbableElements[0];
-      this.currentFocus?.focus({ preventScroll: true });
+    // We want this to scroll the element into view.
+      this.currentFocus?.focus();
       return;
     }
 
@@ -103,9 +112,15 @@ export default class Modal {
     }
 
     this.currentFocus = tabbableElements[currentFocusIndex];
+    // We want this to scroll the element into view.
     this.currentFocus?.focus({ preventScroll: true });
 
-    setTimeout(() => this.checkFocus());
+    setTimeout(() => {
+      this.checkFocus()
+
+      // Make sure to clean up!
+      setTimeout(() => this.tabbableElements = null)
+    });
   };
 
   private handleKeyUp = () => {