+++ src/internal/tabbable.ts
@@ -7,8 +7,26 @@
 // Previously, we used https://www.npmjs.com/package/composed-offset-position, but recursing up an entire node tree took
 // up a lot of CPU cycles and made focus traps unusable in Chrome / Edge.
 //
-function isTakingUpSpace(elem: HTMLElement): boolean {
-  return Boolean(elem.offsetParent || elem.offsetWidth || elem.offsetHeight || elem.getClientRects().length);
+
+
+// Cached compute style calls. This is specifically for browsers that dont support `checkVisibility()`
+const computedStyleMap = new WeakMap()
+
+function isVisible(el: HTMLElement): boolean {
+  // This is the fastest check.
+  if (typeof el.checkVisibility === "function") {
+    return el.checkVisibility({ checkOpacity: false })
+  }
+
+  // Fallback "polyfill"
+  let computedStyle = computedStyleMap.get(el)
+
+  if (!computedStyle) {
+    computedStyle = window.getComputedStyle(el)
+    computedStyleMap.set(el, computedStyle)
+  }
+
+  return computedStyle.visibility !== "hidden" && computedStyle.display !== "none"
 }
 
 /** Determines if the specified element is tabbable using heuristics inspired by https://github.com/focus-trap/tabbable */
@@ -30,14 +48,8 @@ function isTabbable(el: HTMLElement) {
     return false;
   }
 
-  // Elements that are hidden have no offsetParent and are not tabbable
-  if (!isTakingUpSpace(el)) {
-    return false;
-  }
-
-  // Elements without visibility are not tabbable
-  if (window.getComputedStyle(el).visibility === 'hidden') {
-    return false;
+  if (!isVisible(el)) {
+    return false
   }
 
   // Audio and video elements with the controls attribute are tabbable
@@ -73,7 +85,18 @@ export function getTabbableBoundary(root: HTMLElement | ShadowRoot) {
   return { start, end };
 }
 
+/**
+ * This looks funky. Basically a slot's children will always be picked up *if* they're within the `root` element.
+ * However, there is an edge case when, if the `root` is wrapped by another shadow DOM, it won't grab the children.
+ * This fixes that fun edge case.
+ */
+function slotChildrenOutsideRootElement (slotElement: HTMLSlotElement, root: HTMLElement | ShadowRoot) {
+ return (slotElement.getRootNode({ composed: true }) as ShadowRoot | null)?.host !== root;
+}
+
+
 export function getTabbableElements(root: HTMLElement | ShadowRoot) {
+  const walkedEls = new WeakMap()
   const tabbableElements: HTMLElement[] = [];
 
   function walk(el: HTMLElement | ShadowRoot) {
@@ -83,19 +106,16 @@ export function getTabbableElements(root: HTMLElement | ShadowRoot) {
         return;
       }
 
+      if (walkedEls.has(el)) {
+        return
+      }
+      walkedEls.set(el, true)
+
       if (!tabbableElements.includes(el) && isTabbable(el)) {
         tabbableElements.push(el);
       }
 
-      /**
-       * This looks funky. Basically a slot's children will always be picked up *if* they're within the `root` element.
-       * However, there is an edge case when, if the `root` is wrapped by another shadow DOM, it won't grab the children.
-       * This fixes that fun edge case.
-       */
-      const slotChildrenOutsideRootElement = (slotElement: HTMLSlotElement) =>
-        (slotElement.getRootNode({ composed: true }) as ShadowRoot | null)?.host !== root;
-
-      if (el instanceof HTMLSlotElement && slotChildrenOutsideRootElement(el)) {
+      if (el instanceof HTMLSlotElement && slotChildrenOutsideRootElement(el, root)) {
         el.assignedElements({ flatten: true }).forEach((assignedEl: HTMLElement) => {
           walk(assignedEl);
         });
@@ -106,7 +126,7 @@ export function getTabbableElements(root: HTMLElement | ShadowRoot) {
       }
     }
 
-    [...el.children].forEach((e: HTMLElement) => walk(e));
+    for (const e of el.children) { walk(e as HTMLElement) }
   }
 
   // Collect all elements including the root